requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
coverage==6.4.4
Django==4.1.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
Faker==14.2.0
h11==0.13.0
idna==3.4
iniconfig==1.1.1
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
